Add artifact splitting
.meta files are now in YAML format. For a chunk, the .meta file contains a list
of products with an artifact name and a file list for that artifact. Stratum
.meta file also have a list of products, with each entry having an artifact name
and a list of components. The component names in the stratum artifacts reference
the chunk artifacts in the chunk .meta files.

When a stratum is to be split according to a list of artifact names, the
artifact names are looked up in the .meta file for the stratum, and only the
components for matching artifacts are copied to the destination path. The .meta
files for a split stratum should only contain references to files in the split
sub set of the stratum, both for chunk .meta files and for the stratum .meta
file.